+++ include/world/objects/Element.h
@@ -15,6 +15,10 @@ namespace engine::graph {
   class RenderSceneAnalysis;
 }
 
+namespace render {
+  class Object;
+}
+
 class Element : public QObject {
   Q_OBJECT
   Q_PROPERTY(QString id READ id WRITE setId)
@@ -233,6 +237,8 @@ protected:
     result->setName(name().toStdString());
     return result;
   }
+
+  void attachRuntimeAnimationTracks(render::Object& object) const;
 
   void addPendingReference(const QString& property, const QString& id);
   void resolveReferences(const QMap<QString, Element*>& elements);